etl-1/requirements.txt
pymongo==3.11.4
dnspython==2.1.0
PyYAML==5.4